@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:flutter/material.dart';
 import 'package:apple_maps_flutter/apple_maps_flutter.dart';
+import 'package:flutter/material.dart';
 
 import 'page.dart';
 
@@ -17,7 +17,7 @@ class MapUiPage extends ExamplePage {
 
   @override
   Widget build(BuildContext context) {
-    return const MapUiBody();
+    return SafeArea(child: const MapUiBody());
   }
 }
 
@@ -48,7 +48,7 @@ class MapUiBodyState extends State<MapUiBody> {
   bool _pitchGesturesEnabled = true;
   bool _zoomGesturesEnabled = true;
   bool _myLocationEnabled = true;
-  AppleMapController _controller;
+  TrackingMode _trackingMode = TrackingMode.none;
   @override
   void initState() {
     super.initState();
@@ -60,7 +60,7 @@ class MapUiBodyState extends State<MapUiBody> {
   }
 
   Widget _compassToggler() {
-    return FlatButton(
+    return TextButton(
       child: Text('${_compassEnabled ? 'disable' : 'enable'} compass'),
       onPressed: () {
         setState(() {
@@ -71,7 +71,7 @@ class MapUiBodyState extends State<MapUiBody> {
   }
 
   Widget _zoomBoundsToggler() {
-    return FlatButton(
+    return TextButton(
       child: Text(_minMaxZoomPreference.minZoom == null
           ? 'bound zoom'
           : 'release zoom'),
@@ -88,7 +88,7 @@ class MapUiBodyState extends State<MapUiBody> {
   Widget _mapTypeCycler() {
     final MapType nextType =
         MapType.values[(_mapType.index + 1) % MapType.values.length];
-    return FlatButton(
+    return TextButton(
       child: Text('change map type to $nextType'),
       onPressed: () {
         setState(() {
@@ -99,7 +99,7 @@ class MapUiBodyState extends State<MapUiBody> {
   }
 
   Widget _rotateToggler() {
-    return FlatButton(
+    return TextButton(
       child: Text('${_rotateGesturesEnabled ? 'disable' : 'enable'} rotate'),
       onPressed: () {
         setState(() {
@@ -110,7 +110,7 @@ class MapUiBodyState extends State<MapUiBody> {
   }
 
   Widget _scrollToggler() {
-    return FlatButton(
+    return TextButton(
       child: Text('${_scrollGesturesEnabled ? 'disable' : 'enable'} scroll'),
       onPressed: () {
         setState(() {
@@ -121,7 +121,7 @@ class MapUiBodyState extends State<MapUiBody> {
   }
 
   Widget _tiltToggler() {
-    return FlatButton(
+    return TextButton(
       child: Text('${_pitchGesturesEnabled ? 'disable' : 'enable'} tilt'),
       onPressed: () {
         setState(() {
@@ -132,7 +132,7 @@ class MapUiBodyState extends State<MapUiBody> {
   }
 
   Widget _zoomToggler() {
-    return FlatButton(
+    return TextButton(
       child: Text('${_zoomGesturesEnabled ? 'disable' : 'enable'} zoom'),
       onPressed: () {
         setState(() {
@@ -143,7 +143,7 @@ class MapUiBodyState extends State<MapUiBody> {
   }
 
   Widget _myLocationToggler() {
-    return FlatButton(
+    return TextButton(
       child: Text(
           '${_myLocationEnabled ? 'disable' : 'enable'} my location annotation'),
       onPressed: () {
@@ -155,7 +155,7 @@ class MapUiBodyState extends State<MapUiBody> {
   }
 
   Widget _myLocationButtonToggler() {
-    return FlatButton(
+    return TextButton(
       child: Text(
           '${_myLocationButtonEnabled ? 'disable' : 'enable'} my location button'),
       onPressed: () {
@@ -170,6 +170,7 @@ class MapUiBodyState extends State<MapUiBody> {
   Widget build(BuildContext context) {
     final AppleMap appleMap = AppleMap(
       onMapCreated: onMapCreated,
+      trackingMode: _trackingMode,
       initialCameraPosition: _kInitialPosition,
       compassEnabled: _compassEnabled,
       minMaxZoomPreference: _minMaxZoomPreference,
@@ -180,27 +181,20 @@ class MapUiBodyState extends State<MapUiBody> {
       zoomGesturesEnabled: _zoomGesturesEnabled,
       myLocationEnabled: _myLocationEnabled,
       myLocationButtonEnabled: _myLocationButtonEnabled,
+      padding: const EdgeInsets.all(10),
       onCameraMove: _updateCameraPosition,
     );
 
     final List<Widget> columnChildren = <Widget>[
-      Padding(
-        padding: const EdgeInsets.all(10.0),
-        child: Center(
-          child: SizedBox(
-            width: 300.0,
-            height: 200.0,
-            child: appleMap,
-          ),
-        ),
-      ),
+      Expanded(child: appleMap),
     ];
 
     if (_isMapCreated) {
-      columnChildren.add(
-        Expanded(
-          child: ListView(
-            children: <Widget>[
+      columnChildren.addAll([
+        Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Column(
+            children: [
               Text('camera bearing: ${_position.heading}'),
               Text(
                   'camera target: ${_position.target.latitude.toStringAsFixed(4)},'
@@ -208,25 +202,26 @@ class MapUiBodyState extends State<MapUiBody> {
               Text('camera zoom: ${_position.zoom}'),
               Text('camera tilt: ${_position.pitch}'),
               Text(_isMoving ? '(Camera moving)' : '(Camera idle)'),
-              _compassToggler(),
-              _mapTypeCycler(),
-              _zoomBoundsToggler(),
-              _rotateToggler(),
-              _scrollToggler(),
-              _tiltToggler(),
-              _zoomToggler(),
-              _myLocationToggler(),
-              _myLocationButtonToggler(),
+              Wrap(
+                alignment: WrapAlignment.spaceEvenly,
+                children: <Widget>[
+                  _compassToggler(),
+                  _mapTypeCycler(),
+                  _zoomBoundsToggler(),
+                  _rotateToggler(),
+                  _scrollToggler(),
+                  _tiltToggler(),
+                  _zoomToggler(),
+                  _myLocationToggler(),
+                  _myLocationButtonToggler(),
+                ],
+              ),
             ],
           ),
         ),
-      );
+      ]);
     }
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.start,
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: columnChildren,
-    );
+    return Column(children: columnChildren);
   }
 
   void _updateCameraPosition(CameraPosition position) {
@@ -237,7 +232,6 @@ class MapUiBodyState extends State<MapUiBody> {
 
   void onMapCreated(AppleMapController controller) {
     setState(() {
-      _controller = controller;
       _isMapCreated = true;
     });
   }
