@@ -4,10 +4,13 @@
 
 import 'dart:async';
 import 'dart:math';
+import 'dart:typed_data';
 import 'dart:ui';
+import 'dart:ui' as ui;
 
-import 'package:flutter/material.dart';
 import 'package:apple_maps_flutter/apple_maps_flutter.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
 import 'page.dart';
 
@@ -33,10 +36,13 @@ class PlaceAnnotationBodyState extends State<PlaceAnnotationBody> {
   PlaceAnnotationBodyState();
   static final LatLng center = const LatLng(-33.86711, 151.1947171);
 
-  AppleMapController controller;
+  late AppleMapController controller;
   Map<AnnotationId, Annotation> annotations = <AnnotationId, Annotation>{};
-  AnnotationId selectedAnnotation;
+  late AnnotationId selectedAnnotation;
   int _annotationIdCounter = 1;
+  BitmapDescriptor? _annotationIcon;
+  late BitmapDescriptor _iconFromBytes;
+  double _devicePixelRatio = 3.0;
 
   void _onMapCreated(AppleMapController controller) {
     this.controller = controller;
@@ -48,12 +54,12 @@ class PlaceAnnotationBodyState extends State<PlaceAnnotationBody> {
   }
 
   void _onAnnotationTapped(AnnotationId annotationId) {
-    final Annotation tappedAnnotation = annotations[annotationId];
+    final Annotation? tappedAnnotation = annotations[annotationId];
     if (tappedAnnotation != null) {
       setState(() {
-        if (annotations.containsKey(selectedAnnotation)) {
-          final Annotation resetOld = annotations[selectedAnnotation]
-              .copyWith(iconParam: BitmapDescriptor.defaultAnnotation);
+        if (annotations.containsKey(tappedAnnotation)) {
+          final Annotation resetOld =
+              annotations[selectedAnnotation]!.copyWith();
           annotations[selectedAnnotation] = resetOld;
         }
         selectedAnnotation = annotationId;
@@ -76,12 +82,20 @@ class PlaceAnnotationBodyState extends State<PlaceAnnotationBody> {
       annotationId: annotationId,
       icon: iconType == 'marker'
           ? BitmapDescriptor.markerAnnotation
-          : BitmapDescriptor.defaultAnnotation,
+          : iconType == 'pin'
+              ? BitmapDescriptor.defaultAnnotation
+              : iconType == 'customAnnotationFromBytes'
+                  ? _iconFromBytes
+                  : _annotationIcon!,
       position: LatLng(
         center.latitude + sin(_annotationIdCounter * pi / 6.0) / 20.0,
         center.longitude + cos(_annotationIdCounter * pi / 6.0) / 20.0,
       ),
-      infoWindow: InfoWindow(title: annotationIdVal, snippet: '*'),
+      infoWindow: InfoWindow(
+          title: annotationIdVal,
+          anchor: Offset(0.5, 0.0),
+          snippet: '*',
+          onTap: () => print('InfowWindow of id: $annotationId tapped.')),
       onTap: () {
         _onAnnotationTapped(annotationId);
       },
@@ -101,7 +115,7 @@ class PlaceAnnotationBodyState extends State<PlaceAnnotationBody> {
   }
 
   void _changePosition() {
-    final Annotation annotation = annotations[selectedAnnotation];
+    final Annotation annotation = annotations[selectedAnnotation]!;
     final LatLng current = annotation.position;
     final Offset offset = Offset(
       center.latitude - current.latitude,
@@ -118,7 +132,7 @@ class PlaceAnnotationBodyState extends State<PlaceAnnotationBody> {
   }
 
   Future<void> _toggleDraggable() async {
-    final Annotation annotation = annotations[selectedAnnotation];
+    final Annotation annotation = annotations[selectedAnnotation]!;
     setState(() {
       annotations[selectedAnnotation] = annotation.copyWith(
         draggableParam: !annotation.draggable,
@@ -127,8 +141,9 @@ class PlaceAnnotationBodyState extends State<PlaceAnnotationBody> {
   }
 
   Future<void> _changeInfo() async {
-    final Annotation annotation = annotations[selectedAnnotation];
-    final String newSnippet = annotation.infoWindow.snippet + '*';
+    final Annotation annotation = annotations[selectedAnnotation]!;
+    final String newSnippet = annotation.infoWindow.snippet! +
+        (annotation.infoWindow.snippet!.length % 10 == 0 ? '\n' : '*');
     setState(() {
       annotations[selectedAnnotation] = annotation.copyWith(
         infoWindowParam: annotation.infoWindow.copyWith(
@@ -139,7 +154,7 @@ class PlaceAnnotationBodyState extends State<PlaceAnnotationBody> {
   }
 
   Future<void> _changeAlpha() async {
-    final Annotation annotation = annotations[selectedAnnotation];
+    final Annotation annotation = annotations[selectedAnnotation]!;
     final double current = annotation.alpha;
     setState(() {
       annotations[selectedAnnotation] = annotation.copyWith(
@@ -149,7 +164,7 @@ class PlaceAnnotationBodyState extends State<PlaceAnnotationBody> {
   }
 
   Future<void> _toggleVisible() async {
-    final Annotation annotation = annotations[selectedAnnotation];
+    final Annotation annotation = annotations[selectedAnnotation]!;
     setState(() {
       annotations[selectedAnnotation] = annotation.copyWith(
         visibleParam: !annotation.visible,
@@ -157,16 +172,61 @@ class PlaceAnnotationBodyState extends State<PlaceAnnotationBody> {
     });
   }
 
+  Future<void> _createAnnotationImageFromAsset(
+      BuildContext context, double devicelPixelRatio) async {
+    if (_annotationIcon == null) {
+      final ImageConfiguration imageConfiguration =
+          ImageConfiguration(devicePixelRatio: devicelPixelRatio);
+      BitmapDescriptor.fromAssetImage(
+              imageConfiguration, 'assets/red_square.png')
+          .then(_updateBitmap);
+    }
+  }
+
+  void _updateBitmap(BitmapDescriptor bitmap) {
+    setState(() {
+      _annotationIcon = bitmap;
+    });
+  }
+
+  Future<void> _showInfoWindow() async {
+    final Annotation annotation = annotations[selectedAnnotation]!;
+    await this.controller.showMarkerInfoWindow(annotation.annotationId);
+  }
+
+  Future<void> _hideInfoWindow() async {
+    final Annotation annotation = annotations[selectedAnnotation]!;
+    this.controller.hideMarkerInfoWindow(annotation.annotationId);
+  }
+
+  Future<bool> _isInfoWindowShown() async {
+    final Annotation annotation = annotations[selectedAnnotation]!;
+    print(
+        'Is InfowWindow visible: ${await this.controller.isMarkerInfoWindowShown(annotation.annotationId)}');
+    return (await this
+        .controller
+        .isMarkerInfoWindowShown(annotation.annotationId))!;
+  }
+
+  Future<void> _getBytesFromAsset(String path, int width) async {
+    ByteData data = await rootBundle.load(path);
+    ui.Codec codec = await ui.instantiateImageCodec(data.buffer.asUint8List(),
+        targetWidth: width);
+    ui.FrameInfo fi = await codec.getNextFrame();
+    _iconFromBytes = BitmapDescriptor.fromBytes(
+        (await fi.image.toByteData(format: ui.ImageByteFormat.png))!
+            .buffer
+            .asUint8List());
+  }
+
   @override
   Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: <Widget>[
-        Center(
-          child: SizedBox(
-            width: 300.0,
-            height: 200.0,
+    _createAnnotationImageFromAsset(context, _devicePixelRatio);
+    _getBytesFromAsset('assets/red_square.png', 40);
+    return SafeArea(
+      child: Column(
+        children: <Widget>[
+          Expanded(
             child: AppleMap(
               onMapCreated: _onMapCreated,
               initialCameraPosition: const CameraPosition(
@@ -176,61 +236,68 @@ class PlaceAnnotationBodyState extends State<PlaceAnnotationBody> {
               annotations: Set<Annotation>.of(annotations.values),
             ),
           ),
-        ),
-        Expanded(
-          child: SingleChildScrollView(
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: <Widget>[
-                Row(
-                  children: <Widget>[
-                    Column(
-                      children: <Widget>[
-                        FlatButton(
-                          child: const Text('add defaultAnnotation'),
-                          onPressed: () => _add('pin'),
-                        ),
-                        FlatButton(
-                          child: const Text('add markerAnnotation'),
-                          onPressed: () => _add('marker'),
-                        ),
-                        FlatButton(
-                          child: const Text('remove'),
-                          onPressed: _remove,
-                        ),
-                        FlatButton(
-                          child: const Text('change info'),
-                          onPressed: _changeInfo,
-                        ),
-                      ],
-                    ),
-                    Column(
-                      children: <Widget>[
-                        FlatButton(
-                          child: const Text('change alpha'),
-                          onPressed: _changeAlpha,
-                        ),
-                        FlatButton(
-                          child: const Text('toggle draggable'),
-                          onPressed: _toggleDraggable,
-                        ),
-                        FlatButton(
-                          child: const Text('change position'),
-                          onPressed: _changePosition,
-                        ),
-                        FlatButton(
-                          child: const Text('toggle visible'),
-                          onPressed: _toggleVisible,
-                        ),
-                      ],
-                    ),
-                  ],
-                )
+          Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: Wrap(
+              alignment: WrapAlignment.spaceEvenly,
+              children: [
+                TextButton(
+                  child: const Text('add defaultAnnotation'),
+                  onPressed: () => _add('pin'),
+                ),
+                TextButton(
+                  child: const Text('add markerAnnotation'),
+                  onPressed: () => _add('marker'),
+                ),
+                TextButton(
+                  child: const Text('add customAnnotation'),
+                  onPressed: () => _add('customAnnotation'),
+                ),
+                TextButton(
+                  child: const Text('customAnnotation from bytes'),
+                  onPressed: () => _add('customAnnotationFromBytes'),
+                ),
+                TextButton(
+                  child: const Text('remove'),
+                  onPressed: _remove,
+                ),
+                TextButton(
+                  child: const Text('change info'),
+                  onPressed: _changeInfo,
+                ),
+                TextButton(
+                  child: const Text('infoWindow is shown?s'),
+                  onPressed: _isInfoWindowShown,
+                ),
+                TextButton(
+                  child: const Text('change alpha'),
+                  onPressed: _changeAlpha,
+                ),
+                TextButton(
+                  child: const Text('toggle draggable'),
+                  onPressed: _toggleDraggable,
+                ),
+                TextButton(
+                  child: const Text('change position'),
+                  onPressed: _changePosition,
+                ),
+                TextButton(
+                  child: const Text('toggle visible'),
+                  onPressed: _toggleVisible,
+                ),
+                TextButton(
+                  child: const Text('show infoWindow'),
+                  onPressed: _showInfoWindow,
+                ),
+                TextButton(
+                  child: const Text('hide infoWindow'),
+                  onPressed: _hideInfoWindow,
+                ),
               ],
             ),
-          ),
-        ),
-      ],
+          )
+        ],
+      ),
     );
   }
 }
