@@ -2,20 +2,20 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:apple_maps_flutter_example/map_update.dart';
-import 'package:apple_maps_flutter_example/padding.dart';
-import 'package:apple_maps_flutter_example/place_annotation.dart';
-import 'package:apple_maps_flutter_example/place_circle.dart';
-import 'package:apple_maps_flutter_example/place_polyline.dart';
-import 'package:apple_maps_flutter_example/place_polygon.dart';
-import 'package:apple_maps_flutter_example/scrolling_map.dart';
 import 'package:apple_maps_flutter_example/animate_camera.dart';
+import 'package:apple_maps_flutter_example/annotation_icons.dart';
 import 'package:apple_maps_flutter_example/map_click.dart';
 import 'package:apple_maps_flutter_example/map_coordinates.dart';
 import 'package:apple_maps_flutter_example/map_ui.dart';
-import 'package:apple_maps_flutter_example/annotation_icons.dart';
+import 'package:apple_maps_flutter_example/map_update.dart';
 import 'package:apple_maps_flutter_example/move_camera.dart';
+import 'package:apple_maps_flutter_example/padding.dart';
 import 'package:apple_maps_flutter_example/page.dart';
+import 'package:apple_maps_flutter_example/place_annotation.dart';
+import 'package:apple_maps_flutter_example/place_circle.dart';
+import 'package:apple_maps_flutter_example/place_polygon.dart';
+import 'package:apple_maps_flutter_example/place_polyline.dart';
+import 'package:apple_maps_flutter_example/scrolling_map.dart';
 import 'package:flutter/material.dart';
 
 final List<ExamplePage> _allPages = <ExamplePage>[
