@@ -8,7 +8,7 @@
 import Foundation
 
 enum IconType {
-    case PIN, MARKER, CUSTOM
+    case PIN, MARKER, CUSTOM_FROM_ASSET, CUSTOM_FROM_BYTES
 }
 
 class AnnotationIcon: Equatable {
@@ -17,25 +17,39 @@ class AnnotationIcon: Equatable {
     var id: String
     var image: UIImage?
     
-    public init(named name: String, iconType type: IconType? = .PIN, id: String, iconScale: CGFloat? = 1.0) {
-        if type == .CUSTOM {
-            if let uiImage: UIImage =  UIImage.init(named: name) {
-                if let cgImage: CGImage = uiImage.cgImage {
-                    if iconScale != nil && iconScale! - 1 > 0.001 {
-                        let scaledImage: UIImage = UIImage.init(cgImage: cgImage, scale: (iconScale! + 1) * CGFloat(uiImage.scale), orientation: uiImage.imageOrientation)
-                        self.image = scaledImage
-                    }
-                } else {
-                    self.image = uiImage
-                }
-            }
+    public init(id: String, iconType: IconType) {
+        self.iconType = iconType
+        self.id = id
+    }
+    
+    public init(named name: String, id: String, iconScale: CGFloat? = 1.0) {
+        self.iconType = .CUSTOM_FROM_ASSET
+        self.id = id
+        if let uiImage: UIImage =  UIImage.init(named: name) {
+            self.image = self.scaleImage(image: uiImage, scale: iconScale!)
         }
-        self.iconType = type!
+    }
+    
+    public init(fromBytes bytes: FlutterStandardTypedData, id: String) {
+        let screenScale = UIScreen.main.scale
+        let image = UIImage.init(data: bytes.data, scale: screenScale)
+        self.image = image
+        self.iconType = .CUSTOM_FROM_BYTES
         self.id = id
     }
     
     public convenience init() {
-        self.init(named: "", id: "")
+        self.init(id: "", iconType: .PIN)
+    }
+    
+    private func scaleImage(image: UIImage, scale: CGFloat) -> UIImage {
+        guard let cgImage = image.cgImage else {
+            return image
+        }
+        guard abs(scale - 1) >= 0 else {
+            return image
+        }
+        return UIImage.init(cgImage: cgImage, scale: 4.0, orientation: image.imageOrientation)
     }
     
     static func == (lhs: AnnotationIcon, rhs: AnnotationIcon) -> Bool {
