@@ -13,11 +13,14 @@ class FlutterAnnotation: NSObject, MKAnnotation {
     var id :String!
     var title: String?
     var subtitle: String?
+    var infoWindowConsumesTapEvents: Bool = false
     var image: UIImage?
     var alpha: Double?
+    var anchor: Offset = Offset()
     var isDraggable: Bool?
     var wasDragged: Bool = false
     var isVisible: Bool? = true
+    var calloutOffset: Offset = Offset()
     var icon: AnnotationIcon = AnnotationIcon.init()
     
     public init(fromDictionary annotationData: Dictionary<String, Any>, registrar: FlutterPluginRegistrar) {
@@ -28,40 +31,73 @@ class FlutterAnnotation: NSObject, MKAnnotation {
         self.coordinate = CLLocationCoordinate2D(latitude: lat, longitude: long)
         self.title = infoWindow["title"] as? String
         self.subtitle = infoWindow["snippet"] as? String
+        self.infoWindowConsumesTapEvents = infoWindow["consumesTapEvents"] as? Bool ?? false
         self.id = annotationData["annotationId"] as? String
         self.isVisible = annotationData["visible"] as? Bool
         self.isDraggable = annotationData["draggable"] as? Bool
         if let alpha: Double = annotationData["alpha"] as? Double {
             self.alpha = alpha
         }
+        
+        if let anchorJSON: Array<Double> = annotationData["anchor"] as? Array<Double> {
+            self.anchor = Offset(from: anchorJSON)
+        }
+        
         if let iconData: Array<Any> = annotationData["icon"] as? Array<Any> {
             self.icon = FlutterAnnotation.getAnnotationIcon(iconData: iconData, registrar: registrar, annotationId: id)
+        }
+        
+        if let calloutOffsetJSON = infoWindow["anchor"] as? Array<Double> {
+            self.calloutOffset = Offset(from: calloutOffsetJSON)
         }
     }
     
     static private func getAnnotationIcon(iconData: Array<Any>, registrar: FlutterPluginRegistrar, annotationId: String) -> AnnotationIcon {
-        let iconTypeMap: Dictionary<String, IconType> = ["fromAssetImage": IconType.CUSTOM, "defaultAnnotation": IconType.PIN, "markerAnnotation": IconType.MARKER]
+        let iconTypeMap: Dictionary<String, IconType> = ["fromAssetImage": .CUSTOM_FROM_ASSET, "fromBytes": .CUSTOM_FROM_BYTES, "defaultAnnotation": .PIN, "markerAnnotation": .MARKER]
         var icon: AnnotationIcon
         let iconType: IconType = iconTypeMap[iconData[0] as! String] ?? .PIN
         var scaleParam: CGFloat?
-       
-        if iconType == .CUSTOM {
+        
+        if iconType == .CUSTOM_FROM_ASSET {
             let assetPath: String = iconData[1] as! String
             scaleParam = CGFloat(iconData[2] as? Double ?? 1.0)
-            icon = AnnotationIcon(named: registrar.lookupKey(forAsset: assetPath), iconType: iconType, id: annotationId, iconScale: scaleParam)
-        } else {
-            icon = AnnotationIcon(named: "", iconType: iconType, id: annotationId)
+            icon = AnnotationIcon(named: registrar.lookupKey(forAsset: assetPath), id: annotationId, iconScale: scaleParam)
+        } else if iconType == .CUSTOM_FROM_BYTES {
+            icon = AnnotationIcon(fromBytes: iconData[1] as! FlutterStandardTypedData, id: annotationId)
+        }else {
+            icon = AnnotationIcon(id: annotationId, iconType: iconType)
         }
         return icon
     }
     
     static func == (lhs: FlutterAnnotation, rhs: FlutterAnnotation) -> Bool {
-        return  lhs.id == rhs.id && lhs.title == rhs.title && lhs.subtitle == rhs.subtitle && lhs.image == rhs.image && lhs.alpha == rhs.alpha
-            && lhs.isDraggable == rhs.isDraggable && lhs.wasDragged == rhs.wasDragged && lhs.isVisible == rhs.isVisible && lhs.icon == rhs.icon
-            && lhs.coordinate.latitude == rhs.coordinate.latitude && lhs.coordinate.longitude == rhs.coordinate.longitude
+        return lhs.id == rhs.id && lhs.title == rhs.title && lhs.subtitle == rhs.subtitle && lhs.image == rhs.image && lhs.alpha == rhs.alpha && lhs.isDraggable == rhs.isDraggable && lhs.wasDragged == rhs.wasDragged && lhs.isVisible == rhs.isVisible && lhs.icon == rhs.icon && lhs.coordinate.latitude == rhs.coordinate.latitude && lhs.coordinate.longitude == rhs.coordinate.longitude && lhs.infoWindowConsumesTapEvents == rhs.infoWindowConsumesTapEvents && lhs.anchor == rhs.anchor && lhs.calloutOffset == rhs.calloutOffset && lhs.coordinate.latitude == rhs.coordinate.latitude && lhs.coordinate.longitude == rhs.coordinate.longitude
     }
     
     static func != (lhs: FlutterAnnotation, rhs: FlutterAnnotation) -> Bool {
+        return !(lhs == rhs)
+    }
+}
+
+struct Offset {
+    let x: Double
+    let y: Double
+    
+    public init(from json: Array<Double>) {
+        self.x = json[0]
+        self.y = json[1]
+    }
+    
+    public init() {
+        self.x = 0
+        self.y = 0
+    }
+    
+    static func == (lhs: Offset, rhs: Offset) -> Bool {
+        return lhs.x == rhs.x && lhs.y == rhs.y
+    }
+    
+    static func != (lhs: Offset, rhs: Offset) -> Bool {
         return !(lhs == rhs)
     }
 }
