@@ -28,7 +28,6 @@ class BitmapDescriptor {
   /// annotation image. For convenience, there is a predefined set of [AnnotationColors].
   /// See e.g. [AnnotationColor.RED].
   static BitmapDescriptor markerAnnotationWithColor(AnnotationColor color) {
-    assert(color != null);
     return BitmapDescriptor._(<dynamic>['markerAnnotation', color.index]);
   }
 
@@ -36,7 +35,6 @@ class BitmapDescriptor {
   /// annotation image. For convenience, there is a predefined set of [AnnotationColors].
   /// See e.g. [AnnotationColor.RED].
   static BitmapDescriptor defaultAnnotationWithColor(AnnotationColor color) {
-    assert(color != null);
     return BitmapDescriptor._(<dynamic>['defaultAnnotation', color.index]);
   }
 
@@ -49,10 +47,11 @@ class BitmapDescriptor {
   static Future<BitmapDescriptor> fromAssetImage(
     ImageConfiguration configuration,
     String assetName, {
-    AssetBundle bundle,
-    String package,
+    AssetBundle? bundle,
+    String? package,
+    bool mipmaps = true,
   }) async {
-    if (configuration.devicePixelRatio != null) {
+    if (!mipmaps && configuration.devicePixelRatio != null) {
       return BitmapDescriptor._(<dynamic>[
         'fromAssetImage',
         assetName,
@@ -68,6 +67,12 @@ class BitmapDescriptor {
       assetBundleImageKey.name,
       assetBundleImageKey.scale,
     ]);
+  }
+
+  /// Creates a BitmapDescriptor using an array of bytes that must be encoded
+  /// as PNG.
+  static BitmapDescriptor fromBytes(Uint8List byteData) {
+    return BitmapDescriptor._(<dynamic>['fromBytes', byteData]);
   }
 
   final dynamic _json;
