@@ -2,15 +2,15 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:apple_maps_flutter/apple_maps_flutter.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:apple_maps_flutter/apple_maps_flutter.dart';
 
 import 'fake_maps_controllers.dart';
 
-Set<Polygon> _toSet({Polygon p1, Polygon p2, Polygon p3}) {
+Set<Polygon> _toSet({Polygon? p1, Polygon? p2, Polygon? p3}) {
   final Set<Polygon> res = Set<Polygon>.identity();
   if (p1 != null) {
     res.add(p1);
@@ -24,7 +24,7 @@ Set<Polygon> _toSet({Polygon p1, Polygon p2, Polygon p3}) {
   return res;
 }
 
-Widget _mapWithPolygons(Set<Polygon> polygons) {
+Widget _mapWithPolygons(Set<Polygon>? polygons) {
   return Directionality(
     textDirection: TextDirection.ltr,
     child: AppleMap(
@@ -55,13 +55,13 @@ void main() {
     await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1)));
 
     final FakePlatformAppleMap platformAppleMap =
-        fakePlatformViewsController.lastCreatedView;
-    expect(platformAppleMap.polygonsToAdd.length, 1);
+        fakePlatformViewsController.lastCreatedView!;
+    expect(platformAppleMap.polygonsToAdd!.length, 1);
 
-    final Polygon initializedPolygon = platformAppleMap.polygonsToAdd.first;
+    final Polygon initializedPolygon = platformAppleMap.polygonsToAdd!.first;
     expect(initializedPolygon, equals(p1));
-    expect(platformAppleMap.polygonIdsToRemove.isEmpty, true);
-    expect(platformAppleMap.polygonsToChange.isEmpty, true);
+    expect(platformAppleMap.polygonIdsToRemove!.isEmpty, true);
+    expect(platformAppleMap.polygonsToChange!.isEmpty, true);
     debugDefaultTargetPlatformOverride = null;
   });
 
@@ -74,15 +74,15 @@ void main() {
     await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1, p2: p2)));
 
     final FakePlatformAppleMap platformAppleMap =
-        fakePlatformViewsController.lastCreatedView;
-    expect(platformAppleMap.polygonsToAdd.length, 1);
+        fakePlatformViewsController.lastCreatedView!;
+    expect(platformAppleMap.polygonsToAdd!.length, 1);
 
-    final Polygon addedPolygon = platformAppleMap.polygonsToAdd.first;
+    final Polygon addedPolygon = platformAppleMap.polygonsToAdd!.first;
     expect(addedPolygon, equals(p2));
 
-    expect(platformAppleMap.polygonIdsToRemove.isEmpty, true);
+    expect(platformAppleMap.polygonIdsToRemove!.isEmpty, true);
 
-    expect(platformAppleMap.polygonsToChange.isEmpty, true);
+    expect(platformAppleMap.polygonsToChange!.isEmpty, true);
     debugDefaultTargetPlatformOverride = null;
   });
 
@@ -94,12 +94,12 @@ void main() {
     await tester.pumpWidget(_mapWithPolygons(null));
 
     final FakePlatformAppleMap platformAppleMap =
-        fakePlatformViewsController.lastCreatedView;
-    expect(platformAppleMap.polygonIdsToRemove.length, 1);
-    expect(platformAppleMap.polygonIdsToRemove.first, equals(p1.polygonId));
+        fakePlatformViewsController.lastCreatedView!;
+    expect(platformAppleMap.polygonIdsToRemove!.length, 1);
+    expect(platformAppleMap.polygonIdsToRemove!.first, equals(p1.polygonId));
 
-    expect(platformAppleMap.polygonsToChange.isEmpty, true);
-    expect(platformAppleMap.polygonsToAdd.isEmpty, true);
+    expect(platformAppleMap.polygonsToChange!.isEmpty, true);
+    expect(platformAppleMap.polygonsToAdd!.isEmpty, true);
     debugDefaultTargetPlatformOverride = null;
   });
 
@@ -113,12 +113,12 @@ void main() {
     await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p2)));
 
     final FakePlatformAppleMap platformAppleMap =
-        fakePlatformViewsController.lastCreatedView;
-    expect(platformAppleMap.polygonsToChange.length, 1);
-    expect(platformAppleMap.polygonsToChange.first, equals(p2));
+        fakePlatformViewsController.lastCreatedView!;
+    expect(platformAppleMap.polygonsToChange!.length, 1);
+    expect(platformAppleMap.polygonsToChange!.first, equals(p2));
 
-    expect(platformAppleMap.polygonIdsToRemove.isEmpty, true);
-    expect(platformAppleMap.polygonsToAdd.isEmpty, true);
+    expect(platformAppleMap.polygonIdsToRemove!.isEmpty, true);
+    expect(platformAppleMap.polygonsToAdd!.isEmpty, true);
     debugDefaultTargetPlatformOverride = null;
   });
 
@@ -132,10 +132,10 @@ void main() {
     await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p2)));
 
     final FakePlatformAppleMap platformAppleMap =
-        fakePlatformViewsController.lastCreatedView;
-    expect(platformAppleMap.polygonsToChange.length, 1);
+        fakePlatformViewsController.lastCreatedView!;
+    expect(platformAppleMap.polygonsToChange!.length, 1);
 
-    final Polygon update = platformAppleMap.polygonsToChange.first;
+    final Polygon update = platformAppleMap.polygonsToChange!.first;
     expect(update, equals(p2));
     expect(update.visible, false);
     debugDefaultTargetPlatformOverride = null;
@@ -153,12 +153,12 @@ void main() {
     await tester.pumpWidget(_mapWithPolygons(_toSet(p1: p1)));
 
     final FakePlatformAppleMap platformAppleMap =
-        fakePlatformViewsController.lastCreatedView;
-    expect(platformAppleMap.polygonsToChange.length, 1);
-    expect(platformAppleMap.polygonsToChange.first, equals(p1));
+        fakePlatformViewsController.lastCreatedView!;
+    expect(platformAppleMap.polygonsToChange!.length, 1);
+    expect(platformAppleMap.polygonsToChange!.first, equals(p1));
 
-    expect(platformAppleMap.polygonIdsToRemove.isEmpty, true);
-    expect(platformAppleMap.polygonsToAdd.isEmpty, true);
+    expect(platformAppleMap.polygonIdsToRemove!.isEmpty, true);
+    expect(platformAppleMap.polygonsToAdd!.isEmpty, true);
     debugDefaultTargetPlatformOverride = null;
   });
 
@@ -175,11 +175,11 @@ void main() {
     await tester.pumpWidget(_mapWithPolygons(cur));
 
     final FakePlatformAppleMap platformAppleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformAppleMap.polygonsToChange, cur);
-    expect(platformAppleMap.polygonIdsToRemove.isEmpty, true);
-    expect(platformAppleMap.polygonsToAdd.isEmpty, true);
+    expect(platformAppleMap.polygonIdsToRemove!.isEmpty, true);
+    expect(platformAppleMap.polygonsToAdd!.isEmpty, true);
     debugDefaultTargetPlatformOverride = null;
   });
 
@@ -198,15 +198,15 @@ void main() {
     await tester.pumpWidget(_mapWithPolygons(cur));
 
     final FakePlatformAppleMap platformAppleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
-    expect(platformAppleMap.polygonsToChange.length, 1);
-    expect(platformAppleMap.polygonsToAdd.length, 1);
-    expect(platformAppleMap.polygonIdsToRemove.length, 1);
+    expect(platformAppleMap.polygonsToChange!.length, 1);
+    expect(platformAppleMap.polygonsToAdd!.length, 1);
+    expect(platformAppleMap.polygonIdsToRemove!.length, 1);
 
-    expect(platformAppleMap.polygonsToChange.first, equals(p2));
-    expect(platformAppleMap.polygonsToAdd.first, equals(p1));
-    expect(platformAppleMap.polygonIdsToRemove.first, equals(p3.polygonId));
+    expect(platformAppleMap.polygonsToChange!.first, equals(p2));
+    expect(platformAppleMap.polygonsToAdd!.first, equals(p1));
+    expect(platformAppleMap.polygonIdsToRemove!.first, equals(p3.polygonId));
     debugDefaultTargetPlatformOverride = null;
   });
 
@@ -223,11 +223,11 @@ void main() {
     await tester.pumpWidget(_mapWithPolygons(cur));
 
     final FakePlatformAppleMap platformAppleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformAppleMap.polygonsToChange, _toSet(p3: p3));
-    expect(platformAppleMap.polygonIdsToRemove.isEmpty, true);
-    expect(platformAppleMap.polygonsToAdd.isEmpty, true);
+    expect(platformAppleMap.polygonIdsToRemove!.isEmpty, true);
+    expect(platformAppleMap.polygonsToAdd!.isEmpty, true);
     debugDefaultTargetPlatformOverride = null;
   });
 }
