@@ -3,15 +3,15 @@
 // found in the LICENSE file.
 
 import 'package:apple_maps_flutter/apple_maps_flutter.dart';
+import 'package:flutter/foundation.dart'
+    show debugDefaultTargetPlatformOverride;
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter/foundation.dart'
-    show debugDefaultTargetPlatformOverride;
 
 import 'fake_maps_controllers.dart';
 
-Set<Polyline> _toSet({Polyline p1, Polyline p2, Polyline p3}) {
+Set<Polyline> _toSet({Polyline? p1, Polyline? p2, Polyline? p3}) {
   final Set<Polyline> res = Set<Polyline>.identity();
   if (p1 != null) {
     res.add(p1);
@@ -25,7 +25,7 @@ Set<Polyline> _toSet({Polyline p1, Polyline p2, Polyline p3}) {
   return res;
 }
 
-Widget _mapWithPolylines(Set<Polyline> polylines) {
+Widget _mapWithPolylines(Set<Polyline>? polylines) {
   return Directionality(
     textDirection: TextDirection.ltr,
     child: AppleMap(
@@ -54,13 +54,13 @@ void main() {
     await tester.pumpWidget(_mapWithPolylines(_toSet(p1: p1)));
 
     final FakePlatformAppleMap platformAppleMap =
-        fakePlatformViewsController.lastCreatedView;
-    expect(platformAppleMap.polylinesToAdd.length, 1);
+        fakePlatformViewsController.lastCreatedView!;
+    expect(platformAppleMap.polylinesToAdd!.length, 1);
 
-    final Polyline initializedPolyline = platformAppleMap.polylinesToAdd.first;
+    final Polyline initializedPolyline = platformAppleMap.polylinesToAdd!.first;
     expect(initializedPolyline, equals(p1));
-    expect(platformAppleMap.polylineIdsToRemove.isEmpty, true);
-    expect(platformAppleMap.polylinesToChange.isEmpty, true);
+    expect(platformAppleMap.polylineIdsToRemove!.isEmpty, true);
+    expect(platformAppleMap.polylinesToChange!.isEmpty, true);
     debugDefaultTargetPlatformOverride = null;
   });
 
@@ -73,15 +73,15 @@ void main() {
     await tester.pumpWidget(_mapWithPolylines(_toSet(p1: p1, p2: p2)));
 
     final FakePlatformAppleMap platformAppleMap =
-        fakePlatformViewsController.lastCreatedView;
-    expect(platformAppleMap.polylinesToAdd.length, 1);
+        fakePlatformViewsController.lastCreatedView!;
+    expect(platformAppleMap.polylinesToAdd!.length, 1);
 
-    final Polyline addedPolyline = platformAppleMap.polylinesToAdd.first;
+    final Polyline addedPolyline = platformAppleMap.polylinesToAdd!.first;
     expect(addedPolyline, equals(p2));
-    expect(platformAppleMap.polylineIdsToRemove.isEmpty, true);
+    expect(platformAppleMap.polylineIdsToRemove!.isEmpty, true);
 
-    expect(platformAppleMap.polylinesToChange.length, 1);
-    expect(platformAppleMap.polylinesToChange.first, equals(p1));
+    expect(platformAppleMap.polylinesToChange!.length, 1);
+    expect(platformAppleMap.polylinesToChange!.first, equals(p1));
     debugDefaultTargetPlatformOverride = null;
   });
 
@@ -93,12 +93,12 @@ void main() {
     await tester.pumpWidget(_mapWithPolylines(null));
 
     final FakePlatformAppleMap platformAppleMap =
-        fakePlatformViewsController.lastCreatedView;
-    expect(platformAppleMap.polylineIdsToRemove.length, 1);
-    expect(platformAppleMap.polylineIdsToRemove.first, equals(p1.polylineId));
+        fakePlatformViewsController.lastCreatedView!;
+    expect(platformAppleMap.polylineIdsToRemove!.length, 1);
+    expect(platformAppleMap.polylineIdsToRemove!.first, equals(p1.polylineId));
 
-    expect(platformAppleMap.polylinesToChange.isEmpty, true);
-    expect(platformAppleMap.polylinesToAdd.isEmpty, true);
+    expect(platformAppleMap.polylinesToChange!.isEmpty, true);
+    expect(platformAppleMap.polylinesToAdd!.isEmpty, true);
     debugDefaultTargetPlatformOverride = null;
   });
 
@@ -112,12 +112,12 @@ void main() {
     await tester.pumpWidget(_mapWithPolylines(_toSet(p1: p2)));
 
     final FakePlatformAppleMap platformAppleMap =
-        fakePlatformViewsController.lastCreatedView;
-    expect(platformAppleMap.polylinesToChange.length, 1);
-    expect(platformAppleMap.polylinesToChange.first, equals(p2));
+        fakePlatformViewsController.lastCreatedView!;
+    expect(platformAppleMap.polylinesToChange!.length, 1);
+    expect(platformAppleMap.polylinesToChange!.first, equals(p2));
 
-    expect(platformAppleMap.polylineIdsToRemove.isEmpty, true);
-    expect(platformAppleMap.polylinesToAdd.isEmpty, true);
+    expect(platformAppleMap.polylineIdsToRemove!.isEmpty, true);
+    expect(platformAppleMap.polylinesToAdd!.isEmpty, true);
     debugDefaultTargetPlatformOverride = null;
   });
 
@@ -131,10 +131,10 @@ void main() {
     await tester.pumpWidget(_mapWithPolylines(_toSet(p1: p2)));
 
     final FakePlatformAppleMap platformAppleMap =
-        fakePlatformViewsController.lastCreatedView;
-    expect(platformAppleMap.polylinesToChange.length, 1);
+        fakePlatformViewsController.lastCreatedView!;
+    expect(platformAppleMap.polylinesToChange!.length, 1);
 
-    final Polyline update = platformAppleMap.polylinesToChange.first;
+    final Polyline update = platformAppleMap.polylinesToChange!.first;
     expect(update, equals(p2));
     expect(update.visible, true);
     debugDefaultTargetPlatformOverride = null;
@@ -153,11 +153,11 @@ void main() {
     await tester.pumpWidget(_mapWithPolylines(cur));
 
     final FakePlatformAppleMap platformAppleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
     expect(platformAppleMap.polylinesToChange, cur);
-    expect(platformAppleMap.polylineIdsToRemove.isEmpty, true);
-    expect(platformAppleMap.polylinesToAdd.isEmpty, true);
+    expect(platformAppleMap.polylineIdsToRemove!.isEmpty, true);
+    expect(platformAppleMap.polylinesToAdd!.isEmpty, true);
     debugDefaultTargetPlatformOverride = null;
   });
 
@@ -176,15 +176,15 @@ void main() {
     await tester.pumpWidget(_mapWithPolylines(cur));
 
     final FakePlatformAppleMap platformAppleMap =
-        fakePlatformViewsController.lastCreatedView;
+        fakePlatformViewsController.lastCreatedView!;
 
-    expect(platformAppleMap.polylinesToChange.length, 1);
-    expect(platformAppleMap.polylinesToAdd.length, 1);
-    expect(platformAppleMap.polylineIdsToRemove.length, 1);
+    expect(platformAppleMap.polylinesToChange!.length, 1);
+    expect(platformAppleMap.polylinesToAdd!.length, 1);
+    expect(platformAppleMap.polylineIdsToRemove!.length, 1);
 
-    expect(platformAppleMap.polylinesToChange.first, equals(p2));
-    expect(platformAppleMap.polylinesToAdd.first, equals(p1));
-    expect(platformAppleMap.polylineIdsToRemove.first, equals(p3.polylineId));
+    expect(platformAppleMap.polylinesToChange!.first, equals(p2));
+    expect(platformAppleMap.polylinesToAdd!.first, equals(p1));
+    expect(platformAppleMap.polylineIdsToRemove!.first, equals(p3.polylineId));
     debugDefaultTargetPlatformOverride = null;
   });
 
@@ -202,11 +202,11 @@ void main() {
       await tester.pumpWidget(_mapWithPolylines(cur));
 
       final FakePlatformAppleMap platformAppleMap =
-          fakePlatformViewsController.lastCreatedView;
+          fakePlatformViewsController.lastCreatedView!;
 
       expect(platformAppleMap.polylinesToChange, _toSet(p2: p2));
-      expect(platformAppleMap.polylineIdsToRemove.isEmpty, true);
-      expect(platformAppleMap.polylinesToAdd.isEmpty, true);
+      expect(platformAppleMap.polylineIdsToRemove!.isEmpty, true);
+      expect(platformAppleMap.polylinesToAdd!.isEmpty, true);
       debugDefaultTargetPlatformOverride = null;
     },
     skip: true,
